# [Opt] [Refactor] [Bug] Use state machines for optimization on variables