--- a/taichi/analysis/gather_loads_and_stores_in_blocks.cpp
+++ b/taichi/analysis/gather_loads_and_stores_in_blocks.cpp
@@ -6,18 +6,19 @@ TLANG_NAMESPACE_BEGIN
 // Gather all loads, stores and atomic operations for each block.
 class LoadsAndStoresSearcher : public BasicStmtVisitor {
  private:
-  std::unordered_map<Block *, std::vector<Stmt *>> loads_and_stores;
+  std::unique_ptr<std::unordered_map<Block *, std::vector<Stmt *>>> loads_and_stores;
 
  public:
   LoadsAndStoresSearcher() {
     allow_undefined_visitor = true;
     invoke_default_visitor = true;
+    loads_and_stores = std::make_unique<std::unordered_map<Block *, std::vector<Stmt *>>>();
   }
 
   void found(Stmt *store_stmt) {
     for (Block *block = store_stmt->parent; block != nullptr;
          block = block->parent) {
-      stores[block].push_back(store_stmt);
+      (*loads_and_stores)[block].push_back(store_stmt);
     }
   }
 
@@ -37,17 +38,17 @@ class LoadsAndStoresSearcher : public BasicStmtVisitor {
     found(stmt);
   }
 
-  static const std::unordered_map<Block *, std::vector<Stmt *>> &run(
+  static std::unique_ptr<std::unordered_map<Block *, std::vector<Stmt *>>> run(
       IRNode *root) {
     LoadsAndStoresSearcher searcher;
     root->accept(&searcher);
-    return searcher.stores;
+    return std::move(searcher.loads_and_stores);
   }
 };
 
 namespace irpass::analysis {
-const std::unordered_map<Block *, std::vector<Stmt *>>
-    &gather_loads_and_stores_in_blocks(IRNode *root) {
+std::unique_ptr<std::unordered_map<Block *, std::vector<Stmt *>>>
+    gather_loads_and_stores_in_blocks(IRNode *root) {
   return LoadsAndStoresSearcher::run(root);
 }
 }  // namespace irpass::analysis