--- a/taichi/backends/opengl/opengl_api.cpp
+++ b/taichi/backends/opengl/opengl_api.cpp
@@ -247,7 +247,7 @@ bool initialize_opengl(bool error_tolerance) {
   static std::optional<bool> supported;  // std::nullopt
 
   if (supported.has_value()) {  // this function has been called before
-    if (supported.value()) {  // detected to be true in last call
+    if (supported.value()) {    // detected to be true in last call
       return true;
     } else {
       if (!error_tolerance)  // not called from with_opengl