… taichi/analysis/has_load_or_atomic.cpp
… taichi/program/program.cpp
… taichi/transforms/simplify.cpp
… taichi/transforms/variable_optimization.cpp
… tests/python/test_optimization.py