<!-- Thank for your PR! If it's your first time contributing to Taichi, please make sure you have read Contributor Guideline(https://taichi.readthedocs.io/en/latest/contributor_guide.html) (last update: March 26, 2019). -->

<!-- Please always prepend your PR title with tags such as [Metal], [CUDA], [Doc], [Example]. Use a lowercased tag (e.g. [cuda]), for PRs that are invisible to end-users (e.g. intermediate implementation). More details: http://taichi.readthedocs.io/en/latest/contributor_guide.html#pr-title-tags -->

Related issue = #656
Related PR = https://github.com/taichi-dev/taichi/pull/858#issuecomment-618752501

Fix #882 (related PR = #662)
Fix crash when `print_benchmark_stat=True` and there is `/` in `PYTEST_CURRENT_TEST`

Benchmark: **total running time** of 347 tests at commit https://github.com/taichi-dev/taichi/pull/859/commits/d225b63f61d23fc391211771acc8ae6163ee226e.
**before**: (rerun 3 times) 224.65s 225.20s 220.76s (average: 223.54s)
**after**: 218.94s 214.22s 210.33s (average: 214.50s)

After the commit https://github.com/taichi-dev/taichi/pull/859/commits/10ee8454ac3f5940e44717adcfcce9c4ecdcbb06: 213.52s
`master` branch with commit https://github.com/taichi-dev/taichi/pull/859/commits/10ee8454ac3f5940e44717adcfcce9c4ecdcbb06 cherry-picked: 214.64s

Optimization factor is slightly decreased, due to more cautious treatment on loops:
**before**:
![benchmark20200426_4](https://user-images.githubusercontent.com/22582118/80321831-6124f680-87ee-11ea-855b-b85676b9782d.png)
**after**:
![benchmark20200426_3](https://user-images.githubusercontent.com/22582118/80321834-66824100-87ee-11ea-941f-a489491752bc.png)



[[Click here for the format server]](http://kun.csail.mit.edu:31415/)