@xumingkuan 2020-04-24
I need to know if an `AtomicOpStmt` is used in other places quickly (say O(1)) to achieve an O(n * (number of statements to be eliminated)) running time in total in the whole pass. (The `optimize_local_variable` pass is now O(n * (number of statements to be eliminated) * (number of `AllocaStmt`s)))

Is there a better way than passing `std::unordered_set<AtomicOpStmt *> *used_atomics;` to the constructor?