@xumingkuan 2020-04-24
What about making it a static variable, and provide a function to rebuild the set?