@yuanming-hu 2020-04-24
I'm not sure if I fully understand the context of your question, but I believe you can make your own decision :-)