@xumingkuan 2020-04-25
My brain is fried... I found `gather_loads_and_stores_in_blocks` is not convenient to use:
For each loop (O(1)?), for each load/store statement (O(n)), **for each variable that may be modified by that statement** (O(n)), do something like `maybe_store()`. 

This is causing at least O(n^3) time in this pass, and finding variables that may be modified by a statement is non-trivial and can be an overkill.

I use the pattern of simply visiting twice for now. Maybe we want to split `GlobalPtrStmt` out, because we can do aliasing analysis for the other two types of variables easily.