@xumingkuan 2020-04-25
```
kernel {
  <i32 x1> $0 = arg[0]
  <i32*x1> $1 = global ptr [S2place_i32], index [] activate=true
  <gen*x1> $4 = get root
  <i32 x1> $5 = linearized(ind {}, stride {})
  <gen*x1> $6 = [S0root][root]::lookup($4, $5) activate = false
  <gen*x1> $7 = get child [S0root->S1dense] $6
  <i32 x1> $8 = linearized(ind {}, stride {})
  <gen*x1> $9 = [S1dense][dense]::lookup($7, $8) activate = false
  <i32*x1> $10 = get child [S1dense->S2place_i32] $9
  <i32 x1> $11 = shuffle $10[0]
  <i32 x1> $2 : global store [$11 <- $0]
}
```
So `ArgLoadStmt` can also be a variable?