@xumingkuan 2020-04-26
Currently `VariableOptimize` only supports optimization on local variables. I'll add support of global temps/pointers in the next PR.