@xumingkuan 2020-04-27
Shall I rename state_machine.cpp?